+++ CHANGELOG.md
@@ -42,7 +42,9 @@ Breaking Changes:
 * Remove Masterpass Card support
 * Remove Amex Express Checkout Card support
 
-## Unreleased
+## 2.24.0
+* Add * `GatewayRejectionReason.RiskThreshold` to `Transaction`
+* Update @braintree/wrap-promise to v2.1.0
 * Add `networkTransactionId` to `CreditCardVerification`
 * Add `retrievalReferenceNumber` to `Transaction`
 * Add `productSku` to `Transaction`
@@ -59,6 +61,8 @@ Breaking Changes:
   * `RiskData.CustomerLocationZipIsInvalid`
   * `RiskData.CustomerLocationZipIsTooLong`
   * `RiskData.CustomerTenureIsTooLong`
+* Add `processedWithNetworkToken` to `Transaction`
+* Add `isNetworkTokenized` to `CreditCard`
 
 ## 2.23.0
 * Add `threeDSecurePassThru` parameters to `Customer.create`, `PaymentMethod.create`, `CreditCard.create`, `Customer.update`, `PaymentMethod.update` and `CreditCard.update`